refactor(navbar): rearrange components, add stories
