+++ src/components/NavBar/primitives.tsx
@@ -1,8 +1,28 @@
 import React from 'react';
-import { Eye, Grid, Home, Search } from 'react-feather';
 import { Link } from 'react-router-dom';
+import { Home, Search, Eye, Grid } from 'react-feather';
+import { Styles, mediaQueries, Colors, iconStyle } from '../../constants/style';
 
-import { Colors, iconStyle, mediaQueries, Styles } from '../../constants/style';
+/*tslint:disable: no-require-imports no-var-requires */
+const sbLogo = require('@sbac/sbac-ui-kit/src/images/SmarterBalanced-Logo.png') as string;
+
+export const HeaderLogo: React.SFC = () => (
+  <a href="https://www.smarterbalanced.org/" target="_blank" rel="noopener noreferrer">
+    <img alt="Smarter Balanced logo" src={sbLogo} className="sb-logo" />
+    <style jsx>{`
+      .sb-logo {
+        width: 128px;
+        padding: 3px ${Styles.paddingUnit} 3px 3px;
+      }
+
+      @media ${mediaQueries.tabletAndMobile} {
+        .sb-logo {
+          width: 100px;
+        }
+      }
+    `}</style>
+  </a>
+);
 
 export interface MenuItemProps {
   name: string;
